Separate stress and fatigue computation
The mooring_fatigue processor was taking measurements of strain on a
chain and computing fatigue damage. Modifying it to compute the fatigue
from stress allows the fatigue computation to be reused for other types
of structures. The function computing stress in the material induced
from strain on the cahain is moved to a seperate processor.
