+++ camille/process/__init__.py
@@ -13,8 +13,9 @@ Available processors
 * :func:`~camille.process.low_pass`
 * :func:`~camille.process.high_pass`
 * :func:`~camille.process.band_pass`
-* :func:`~camille.process.mooring_fatigue`
+* :func:`~camille.process.fatigue`
 * :func:`~camille.process.rolling_median`
+* :func:`~camille.process.mooring_stress`
 """
 
 from .atm_stb import process as atm_stb
@@ -24,8 +25,9 @@ from .lidar import process as lidar
 from .pass_filter import low_pass
 from .pass_filter import high_pass
 from .pass_filter import band_pass
-from .mooring_fatigue import process as mooring_fatigue
+from .fatigue import process as fatigue
 from .rolling_median import process as rolling_median
+from .mooring_stress import process as mooring_stress
 
 __all__ = [
     'atm_stb',
@@ -35,6 +37,7 @@ __all__ = [
     'low_pass',
     'high_pass',
     'band_pass',
-    'mooring_fatigue',
-    'rolling_median'
+    'fatigue',
+    'rolling_median',
+    'mooring_stress'
 ]
